Checking Nodes in Cluster

+++ app/models/foreman_proxmox/proxmoxserver.rb
@@ -8,12 +8,88 @@ module ForemanProxmox
     def setup_httpclient
       @client= HTTPClient.new
       @client.ssl_config.verify_mode= OpenSSL::SSL::VERIFY_NONE
+      @node = get_first_node_in_cluster
       $LOG= Logger.new("/tmp/proxmox_debug.log")
       $LOG.error("Created HttpClient")
     end
     
-    def authenticate_client
+    def get_first_node_in_cluster
+      if !check_ip_connectivity then
+        flash[:error] = "Proxmoxserver seems down, try again"
+        return nil
+      end
+      authenticate_client
+      nodes_response = client.get("https://#{self.ip}:8006/api2/json/nodes")
+      nodes = JSON.parse(nodes_response.body)
+      return nodes["data"][0]["node"]
+    end
+    
+    def check_ip_connectivity
       if @client == nil then setup_httpclient end
+      code_response = @client.get("https://#{self.ip}:8006/api2/json/access/ticket")
+      if code_response != 200 then
+        self.ip = Proxmoxserver.find(self.id + 1)
+        self.save
+        return false
+      end
+      return true
+    end
+    
+    def find_node_for_vmid(vmid)
+      authenticate_client
+      nodes_response = client.get("https://#{self.ip}:8006/api2/json/nodes")
+      nodes = JSON.parse(nodes_response.body)
+      current_node_id = 0
+      current_node = nodes["data"][current_node]
+      while current_node != nil
+        node_name = nodes["data"][current_node_id]["node"]
+        current_vm_id = 0
+        vms_response = client.get("https://#{self.ip}:8006/api2/json/nodes/#{node_name}/qemu")
+        vms = JSON.parse(vms_response.body)
+        current_vm = vms["data"][current_vm_id]
+        while cur_vm != nil do
+          if vms["data"][current_vm_id]["vmid"] == vmid
+            @node = node_name
+          end
+          current_vm_id+=1
+          current_vm = vms["data"][current_vm_id]
+        end
+        current_node_id+=1
+        current_node = nodes["data"][current_node_id]
+      end
+    end
+    
+    def get_next_free_vmid
+      authenticate_client
+      nodes_response = client.get("https://#{self.ip}:8006/api2/json/nodes")
+      nodes = JSON.parse(nodes_response.body)
+      current_node_id = 0
+      highest_vmid = 0
+      current_node = nodes["data"][current_node_id]
+      while current_node != nil 
+        node_name = nodes["data"][current_node]["node"]
+        current_vm_id = 0
+        vms_response = client.get("https://#{self.ip}:8006/api2/json/nodes/#{node_name}/qemu")
+        vms = JSON.parse(vms_response)
+        current_vm = vms["data"][current_vm_id]
+        while current_vm != nil do
+          if vms["data"][current_vm_id]["vmid"].to_i > highest_vmid 
+            highest_vmid = vms["data"][current_vm_id]["vmid"].to_i
+          end
+          current_vm_id+=1
+          current_vm = vms["data"][current_vm_id]
+        end 
+        current_node_id+=1
+        current_node = nodes["data"][current_node_id]
+      end
+      return highest_vmid+1
+    end
+    
+    def authenticate_client
+      if !check_ip_connectivity then
+        flash[:error] = "Proxmoxserver seems down, try again"
+        return nil
+      end
       domain= "https://#{self.ip}:8006/"
       url= URI.parse(domain)
       credentials= {:username => "#{self.username}@pam", :password => self.password}
@@ -33,17 +109,25 @@ module ForemanProxmox
     
     #manage kvms
     def create_ide(vmid, size)
+      if !check_ip_connectivity then
+        flash[:error] = "Proxmoxserver seems down, try again"
+        return nil
+      end
       authenticate_client
       body= { :filename => "vm-#{vmid}-disk-0.qcow2", :format => "qcow2", :size => size, :vmid => vmid}
-      testres= @client.post("https://#{self.ip}:8006/api2/json/nodes/proxmox/storage/local/content",body,@header)
+      testres= @client.post("https://#{self.ip}:8006/api2/json/nodes/#{@node}/storage/local/content",body,@header)
       $LOG.error("Body: #{testres.body}")
       $LOG.error("Header: #{testres.header}")
     end
     
     def create_kvm(vmid, sockets, cores ,memory,mac)
+      if !check_ip_connectivity then
+        flash[:error] = "Proxmoxserver seems down, try again"
+        return nil
+      end
       authenticate_client
       body= { :vmid => vmid, :sockets => sockets, :cores => cores, :memory => memory, :net0 => "e1000=#{mac},bridge=vmbr1", :ide0 => "volume=local:#{vmid}/vm-#{vmid}-disk-0.qcow2,media=disk"}
-      testres= @client.post("https://#{self.ip}:8006/api2/json/nodes/proxmox/qemu",body,@header)
+      testres= @client.post("https://#{self.ip}:8006/api2/json/nodes/#{@node}/qemu",body,@header)
       $LOG.error("Body: #{testres.body}")
       $LOG.error("Header: #{testres.header}")
     end
@@ -52,28 +136,44 @@ module ForemanProxmox
     end
     
     def delete_kvm(vmid)
-      authenticate_client
+      if !check_ip_connectivity then
+        flash[:error] = "Proxmoxserver seems down, try again"
+        return nil
+      end
+      find_node_for_vmid(vmid)
       testres= @client.delete("https://#{self.ip}:8006/api2/json/nodes/proxmox/qemu/#{vmid}",{},@header)
       $LOG.error("Body: #{testres.body}")
       $LOG.error("Header: #{testres.header}")
     end
     
     def start_kvm(vmid)
-      authenticate_client
+      if !check_ip_connectivity then
+        flash[:error] = "Proxmoxserver seems down, try again"
+        return nil
+      end
+      find_node_for_vmid(vmid)
       testres= @client.post("https://#{self.ip}:8006/api2/json/nodes/proxmox/qemu/#{vmid}/status/start",{},@header)
       $LOG.error("Body: #{testres.body}")
       $LOG.error("Header: #{testres.header}")
     end
     
     def stop_kvm(vmid)
-      authenticate_client
+      if !check_ip_connectivity then
+        flash[:error] = "Proxmoxserver seems down, try again"
+        return nil
+      end
+      find_node_for_vmid(vmid)
       testres= @client.post("https://#{self.ip}:8006/api2/json/nodes/proxmox/qemu/#{vmid}/status/stop",{},@header)
       $LOG.error("Body: #{testres.body}")
       $LOG.error("Header: #{testres.header}")
     end
     
     def reboot_kvm(vmid)
-      authenticate_client
+      if !check_ip_connectivity then
+        flash[:error] = "Proxmoxserver seems down, try again"
+        return nil
+      end
+      find_node_for_vmid(vmid)
       testres= @client.post("https://#{self.ip}:8006/api2/json/nodes/proxmox/qemu/#{vmid}/status/reset",{},@header)
       $LOG.error("Body: #{testres.body}")
       $LOG.error("Header: #{testres.header}")
